+++ CHANGELOG.md
@@ -13,6 +13,11 @@ All changes in this project will be noted in this file.
   - Simply set the protocol version you want to use in the config file, env vars or pass it as a CLI
     argument
   - Even faster implementation, even for Skyhash 1.0
+- New query language: BlueQL
+  - `create keyspace` is now `create space`
+  - `create table` is now `create model`
+  - Similary, all `inspect` queries have been changed
+  - Entities are now of the form `space.model` instead of `ks:tbl`
 
 ## Version 0.7.5